rename: Move domain files about paging
